backend/requirements.txt
matplotlib==3.3.1
numpy==1.20.1
opencv_python_headless==4.2.0.34
imutils==0.5.4
facenet_pytorch==2.5.2
google-cloud-storage==1.43.0
python-firebase==3.0.1
Pillow==8.4.0
protobuf==3.19.1
PyQt5==5.15.6
scikit_learn==1.0.1
torch==1.10.1
torchvision==0.11.2
